+++ resources/views/layouts/navigation.blade.php
@@ -19,6 +19,9 @@
                         <x-nav-link :href="route('tasks')" :active="request()->routeIs('tasks')">
                             {{ __('Tasks') }}
                         </x-nav-link>
+                        <x-nav-link :href="route('releasing')" :active="request()->routeIs('releasing')">
+                            {{ __('Releasing') }}
+                        </x-nav-link>
                     @endif
                     @if (Auth::user()->role === 'admin')
                         <x-nav-link :href="route('integrity-monitor')" :active="request()->routeIs('integrity-monitor')">
@@ -86,6 +89,9 @@
             @if (Auth::user()->role === 'officer')
                 <x-responsive-nav-link :href="route('tasks')" :active="request()->routeIs('tasks')">
                     {{ __('Tasks') }}
+                </x-responsive-nav-link>
+                <x-responsive-nav-link :href="route('releasing')" :active="request()->routeIs('releasing')">
+                    {{ __('Releasing') }}
                 </x-responsive-nav-link>
             @endif
             @if (Auth::user()->role === 'admin')